added a class to hide links in navbar
