+++ .env
@@ -1,0 +1,4 @@
+DBUSER="root"
+DBPASSWORD="mysqlpassword"
+DATABASE="login_db"
+DBHOST="127.0.0.1"
