+++ client/.eslintcache
@@ -1,1 +1,1 @@
-
+[{"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Alacarte.js":"1","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Favorite.js":"2","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Menu.js":"3","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\index.js":"4","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Signup.js":"5","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Home.js":"6","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\components\\Nav\\index.js":"7","C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\utils\\GlobalStore.js":"8"},{"size":269,"mtime":1606763213704,"results":"9","hashOfConfig":"10"},{"size":0,"mtime":1606763196519,"results":"11","hashOfConfig":"10"},{"size":13620,"mtime":1606840388255,"results":"12","hashOfConfig":"10"},{"size":664,"mtime":1606843701627,"results":"13","hashOfConfig":"10"},{"size":2650,"mtime":1606763196523,"results":"14","hashOfConfig":"10"},{"size":2939,"mtime":1606840388254,"results":"15","hashOfConfig":"10"},{"size":3494,"mtime":1606848798030,"results":"16","hashOfConfig":"10"},{"size":1503,"mtime":1606763196524,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"138bgdj",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Alacarte.js",[],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Favorite.js",[],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Menu.js",[],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\index.js",[],["35","36"],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Signup.js",[],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\pages\\Home.js",[],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\components\\Nav\\index.js",[],"C:\\Users\\rsuar\\Desktop\\sushi-junai\\client\\src\\utils\\GlobalStore.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
