+++ client/src/App.js
@@ -7,6 +7,8 @@ import Favorite from "./pages/Favorite";
 import OrderSummary from "./pages/OrderSummary";
 import Signup from "./pages/Signup";
 import Choice from "./pages/Choice";
+import Login from "./pages/Login";
+ 
 
 
 function App() {
@@ -18,7 +20,8 @@ function App() {
             <Route exact path="/menu" component={Menu}/>
             <Route exact path="/choice" component={Choice}/>
             <Route exact path="/favorite" component={Favorite}/>
-            <Route exact path="/ordersummary" component={OrderSummary}/> 
+            <Route exact path="/ordersummary" component={OrderSummary}/>
+            <Route exact path="/login" component={Login}/>
             <Route exact path="/SignUp" component={Signup}/>
         </Router>
     );
