+++ client/src/components/Nav/style.css
@@ -1,0 +1,3 @@
+.hide {
+    display: none;
+}
