+++ client/src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import { StoreProvider } from './utils/GlobalStore';
 import * as serviceWorker from './serviceWorker';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 ReactDOM.render(
     <React.StrictMode>
