+++ config/config.json
@@ -1,14 +1,14 @@
 {
   "development": {
     "username": "root",
-    "password": "DBPASSWORD",
+    "password": "mysqlpassword",
     "database": "login_db",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
   "test": {
     "username": "root",
-    "password": null,
+    "password": "mysqlpassword",
     "database": "login_db",
     "host": "127.0.0.1",
     "dialect": "mysql"
